Move dom update function to domUpdates file
